backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
boto3>=1.34.0
pydantic==2.5.0
email-validator==2.1.0
bcrypt==4.0.1

# AI/LLM Dependencies
google-generativeai>=0.8.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
